// madnessfeed-client/src/app/types.ts
export interface LoginRequest {
    name: string
    password: string
    // second factor, only one of them is used
    mfaTotp?: string
    mfaPasskey?: string
    mfaResetCode?: string
}

export interface MfaLoginRequest {
    name: string
    password: string
}

export interface MfaLoginOptions {
    totp: boolean
    passkey: boolean
    passkeyChallenge?: string
    passkeyCredentialIds: string[]
}